+++ tests_integration/spcs/test_compute_pool.py
@@ -41,7 +41,7 @@ def test_compute_pool(_test_steps: Tuple[ComputePoolTestSteps, str]):
 
 
 @pytest.mark.integration
-def test_compute_pool_create_from_project_definition(
+def test_compute_pool_deploy_from_project_definition(
     _test_steps: Tuple[ComputePoolTestSteps, str],
     project_directory,
     alter_snowflake_yml,
@@ -55,6 +55,26 @@ def test_compute_pool_create_from_project_definition(
         )
 
         test_steps.create_compute_pool_from_project_definition(compute_pool_name)
+
+        alter_snowflake_yml(
+            "snowflake.yml",
+            "entities.compute_pool",
+            {
+                "type": "compute-pool",
+                "identifier": {
+                    "name": compute_pool_name,
+                },
+                "min_nodes": 1,
+                "max_nodes": 2,
+                "auto_resume": True,
+                "auto_suspend_seconds": 10,
+                "comment": "Upgraded compute pool",
+                "tags": [
+                    {"name": "new_tag", "value": "new_value"},
+                ],
+            },
+        )
+        test_steps.upgrade_compute_pool_from_project_definition()
 
 
 @pytest.fixture