+++ tests_integration/spcs/test_services.py
@@ -57,19 +57,40 @@ def test_service_create_from_project_definition(
     _test_steps: Tuple[SnowparkServicesTestSteps, str],
     alter_snowflake_yml,
     project_directory,
-    test_database,
 ):
     test_steps, service_name = _test_steps
+    stage = f"{service_name}_stage"
 
     with project_directory("spcs_service"):
-        alter_snowflake_yml(
-            "snowflake.yml", "entities.service.stage", f"{service_name}_stage"
-        )
+        alter_snowflake_yml("snowflake.yml", "entities.service.stage", stage)
         alter_snowflake_yml(
             "snowflake.yml", "entities.service.identifier.name", service_name
         )
 
         test_steps.deploy_service(service_name)
+
+        alter_snowflake_yml(
+            "snowflake.yml",
+            "entities.service",
+            {
+                "type": "service",
+                "identifier": {
+                    "name": service_name,
+                },
+                "stage": f"{stage}_upgrade",
+                "compute_pool": "snowcli_compute_pool",
+                "spec_file": "spec_upgrade.yml",
+                "min_instances": 1,
+                "max_instances": 2,
+                "query_warehouse": "xsmall",
+                "comment": "Upgraded service",
+                "artifacts": ["spec_upgrade.yml"],
+            },
+        )
+        test_steps.upgrade_service()
+        test_steps.describe_should_return_service(
+            service_name, expected_values={"comment": "Upgraded service"}
+        )
 
 
 @pytest.mark.integration