+++ tests_integration/spcs/testing_utils/spcs_services_utils.py
@@ -20,7 +20,7 @@ import re
 from datetime import datetime
 from pathlib import Path
 from textwrap import dedent
-from typing import List
+from typing import List, Dict
 
 import pytest
 from snowflake.connector import SnowflakeConnection
@@ -89,6 +89,19 @@ class SnowparkServicesTestSteps:
             result, {"status": f"Service {service_name.upper()} successfully created."}
         )
 
+    def upgrade_service(self) -> None:
+        result = self._setup.runner.invoke_with_connection_json(
+            [
+                "spcs",
+                "service",
+                "deploy",
+                "--upgrade",
+            ],
+        )
+        assert_that_result_is_successful_and_output_json_equals(
+            result, {"status": f"Statement executed successfully."}
+        )
+
     def execute_job_service(self, job_service_name: str) -> None:
         result = self._setup.runner.invoke_with_connection_json(
             [
@@ -147,10 +160,14 @@ class SnowparkServicesTestSteps:
         result = self._execute_list()
         assert not_contains_row_with(result.json, {"name": service_name.upper()})
 
-    def describe_should_return_service(self, service_name: str) -> None:
+    def describe_should_return_service(
+        self, service_name: str, expected_values: Dict[str, str] = {}
+    ) -> None:
         result = self._execute_describe(service_name)
         assert result.json
         assert result.json[0]["name"] == service_name.upper()  # type: ignore
+        for key, value in expected_values.items():
+            assert result.json[0][key] == value
 
     def set_unset_service_property(self, service_name: str) -> None:
         comment = "test comment"